feature/backend/add tasks structure /WIN
