+++ backend/controllers/tasks/deleteOperations/index.js
@@ -1,0 +1,11 @@
+const deleteTask = require('./deleteTask');
+const getDeletedTasks = require('./getDeletedTasks');
+const restoreTask = require('./restoreTask');
+const permanentDeleteTask = require('./permanentDeleteTask');
+
+module.exports = {
+  deleteTask,
+  getDeletedTasks,
+  restoreTask,
+  permanentDeleteTask,
+};
