+++ backend/package.json
@@ -5,7 +5,7 @@
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
-    "dev": "nodemon server.js",    
+    "dev": "nodemon server.js",
     "test": "cross-env NODE_ENV=test jest",
     "postinstall": "prisma generate"
   },
@@ -30,6 +30,6 @@
     "cross-env": "^7.0.3",
     "jest": "^29.7.0",
     "supertest": "^7.0.0",
-    "prisma": "^6.5.0"    
+    "prisma": "^6.5.0"
   }
 }
