+++ frontend/package.json
@@ -36,6 +36,6 @@
     "vite": "^6.2.0",
     "vitest": "^3.1.1",
     "prop-types": "^15.8.1",
-    "dotenv": "^16.4.7"    
+    "dotenv": "^16.4.7"
   }
 }
