+++ frontend/src/index.css
@@ -3,8 +3,6 @@
 @tailwind components;
 @tailwind utilities;
 
-
-
 /* Temporary force test 
 .bg-green-500 {
   background-color: #10b981 !important;
