+++ frontend/tailwind.config.js
@@ -1,8 +1,5 @@
 export default {
-  content: [
-    "./index.html",
-    "./src/**/*.{js,ts,jsx,tsx}",
-  ],
+  content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}'],
   important: true, // Force !important on all utilities
   corePlugins: {
     preflight: false, // Disable default styles
@@ -11,4 +8,4 @@ export default {
     extend: {},
   },
   plugins: [],
-}
+};
